+++ changelog.rst
@@ -3,8 +3,9 @@
 Change Log
 ##########
 
-- In development
+- Version 0.7 (2024-07-07)
 
+  - Support custom file-system locations for undo data via ``undo-fu-session-make-file-name-function``.
   - Support derived major-modes when checking ``undo-fu-session-incompatible-major-modes``.
   - Add ``undo-fu-session-temp-directories`` so additional directories can be considered temporary.